decouple SideNav from default layout
